# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Electronics Retail ERP & POS System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products with optional search/filter
      parameters:
        - name: search
          in: query
          schema:
            type: ["string", "null"]
        - name: categoryId
          in: query
          schema:
            type: ["integer", "null"]
        - name: lowStock
          in: query
          schema:
            type: ["boolean", "null"]
        - name: isActive
          in: query
          schema:
            type: ["boolean", "null"]
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a new product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductBody"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Soft-delete product (marks inactive)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product deactivated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /products/barcode/{barcode}:
    get:
      operationId: getProductByBarcode
      tags: [products]
      summary: Look up product by barcode (for POS scanning)
      parameters:
        - name: barcode
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found

  # ─── INVENTORY / STOCK MOVEMENTS ─────────────────────────────────────────────
  /inventory/movements:
    get:
      operationId: listStockMovements
      tags: [inventory]
      summary: List stock movements
      parameters:
        - name: productId
          in: query
          schema:
            type: ["integer", "null"]
        - name: movementType
          in: query
          schema:
            type: ["string", "null"]
        - name: fromDate
          in: query
          schema:
            type: ["string", "null"]
        - name: toDate
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of stock movements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StockMovement"
    post:
      operationId: createStockMovement
      tags: [inventory]
      summary: Create a stock movement (purchase/adjustment/return)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStockMovementBody"
      responses:
        "201":
          description: Movement recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockMovement"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]

    # Auth
...
    Product:
      type: object
      properties:
        id:
          type: integer
        sku:
          type: string
        barcode:
          type: ["string", "null"]
        name:
          type: string
        description:
          type: ["string", "null"]
        categoryId:
          type: integer
        categoryName:
          type: string
        brand:
          type: ["string", "null"]
        model:
          type: ["string", "null"]
        costPrice:
          type: number
        sellingPrice:
          type: number
        gstRate:
          type: number
        currentStock:
          type: integer
        reorderLevel:
          type: integer
        unit:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
      required: [id, sku, name, categoryId, categoryName, costPrice, sellingPrice, gstRate, currentStock, reorderLevel, unit, isActive, createdAt]
    CreateProductBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        categoryId:
          type: integer
        brand:
          type: ["string", "null"]
        model:
          type: ["string", "null"]
        barcode:
          type: ["string", "null"]
        costPrice:
          type: number
        sellingPrice:
          type: number
        gstRate:
          type: number
        reorderLevel:
          type: integer
        unit:
          type: string
        initialStock:
          type: integer
      required: [name, categoryId, costPrice, sellingPrice, gstRate, reorderLevel, unit, initialStock]
    UpdateProductBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        categoryId:
          type: ["integer", "null"]
        brand:
          type: ["string", "null"]
        model:
          type: ["string", "null"]
        barcode:
          type: ["string", "null"]
        costPrice:
          type: ["number", "null"]
        sellingPrice:
          type: ["number", "null"]
        gstRate:
          type: ["number", "null"]
        reorderLevel:
          type: ["integer", "null"]
        unit:
          type: ["string", "null"]
        isActive:
          type: ["boolean", "null"]

    # Stock Movements
    StockMovement:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        productSku:
          type: string
        movementType:
          type: string
          enum: [PURCHASE, SALE, RETURN, ADJUSTMENT]
        quantity:
          type: integer
        previousStock:
          type: integer
        newStock:
          type: integer
        reference:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdBy:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, productId, productName, productSku, movementType, quantity, previousStock, newStock, createdAt]
    CreateStockMovementBody:
      type: object
      properties:
        productId:
          type: integer
        movementType:
          type: string
          enum: [PURCHASE, ADJUSTMENT, RETURN]
        quantity:
          type: integer
        reference:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
      required: [productId, movementType, quantity]

    # Sales
...